reestructura de parcelas y campos nuevos
Se agregó la funcionalidad de eliminar en niveles de administración. Se
agregaron los atributos de institución que acredita (tecnicos),
propiedad (grales. de parcela), terminó el ciclo (caracterización),
datos de riego disponibles: a)tiempo y gasto, b)fecha y lámina, c)fecha
y volumen (riegos). Se reestructuró el módulo de parcelas a manera de
pestañas. Se guardó en sesión el usuario loggeado.

+++ src/main/java/com/saprigrat/data/InterfazInsercion.java
@@ -83,8 +83,8 @@ public class InterfazInsercion
 								if(valor instanceof Double)
 									cs.setBigDecimal(i + 2, new BigDecimal((Double)valor));
 								else
-									if(valor instanceof Long[])
-										cs.setArray(i + 2, con.createArrayOf("decimal", (Long[]) valor));
+									if(valor instanceof Double[])
+										cs.setArray(i + 2, con.createArrayOf("decimal", (Double[]) valor));
 									else
 										if(valor instanceof Integer[])
 											cs.setArray(i + 2, con.createArrayOf("int", (Integer[]) valor));
